refactor:product detail page and product page location
